requirements.txt
Django<2
djangorestframework==3.6.3
djoser==0.6.0
django-taggit==1.3.0
Pillow==7.2.0
sorl-thumbnail==12.6.3

#MySQL-python==1.2.5
mysqlclient==1.3.12